Use application factory to create apps on the fly
